+++ src/main/java/com/paneedah/weaponlib/vehicle/smoothlib/QPTI.java
@@ -7,11 +7,11 @@ import static com.paneedah.mwc.proxies.ClientProxy.MC;
 public class QPTI {
 
     public static float pti(float a, float b) {
-        return (float) InterpolationKit.interpolateValue((double) a, (double) b, (double) MC.getRenderPartialTicks());
+        return (float) InterpolationKit.interpolateValue(a, b, MC.getRenderPartialTicks());
     }
 
     public static double pti(double a, double b) {
-        return InterpolationKit.interpolateValue((double) a, (double) b, (double) MC.getRenderPartialTicks());
+        return InterpolationKit.interpolateValue(a, b, MC.getRenderPartialTicks());
 
     }
